Moves Clusters to transform.py and removes a huge amount of dead code no longer used by the new ingest process.
